bug fix when update user (#88)

+++ client/src/components/main/navbar/navbar.component.js
@@ -121,8 +121,7 @@ function AccountMenu({ handleSubmit, formValues, next_role, userDetails, navbar_
       </h4>
     </StyledMenuItem>
   )
-
-  const httpPostRequestToUpdateUser = (formValues) => {
+  const httpPostRequestToUpdateUser = (formValues, was_teacher, was_student) => {
     let user = {
       _id: formValues._id, password: formValues.password
     }
@@ -137,7 +136,7 @@ function AccountMenu({ handleSubmit, formValues, next_role, userDetails, navbar_
       if (res) {
         axios.post(get_mongo_api(`users/update/${formValues._id}`), formValues).then((response) => {
           if (response.data.success) {
-            alert("המשתמש עודכן");
+            alert("המשתמש עודכן, השינויים ייכנסו לתוקף בהתחברות מחדש למשתמש");
             setUpdated(true)
           } else {
             alert(":בעיה בעדכון" ,response.data.message)
@@ -150,7 +149,14 @@ function AccountMenu({ handleSubmit, formValues, next_role, userDetails, navbar_
               alert(res.data.message)
             }
           })
-        } else {
+        } else if (was_teacher && formValues.isTeacher){
+          let name = `${formValues.first_name} ${formValues.last_name}`
+          axios.post(get_mongo_api(`teachers/update/name`), {_id : formValues._id , name }).then(res => {
+            if (!res.data.success) {
+              alert(res.data.message)
+            }
+          })
+        } else if (was_teacher && !formValues.isTeacher){
           // if the teacher doesn't exist there will be a fuilure - again, nothing will be changed
           axios.delete(get_mongo_api(`teachers/${formValues._id}`)).then(res => {
             if (!res.data.success) {
@@ -158,14 +164,21 @@ function AccountMenu({ handleSubmit, formValues, next_role, userDetails, navbar_
             }
           })
         }
-        if (formValues.isStudent) {
+        if (!was_student && formValues.isStudent) {
           // same as the teacher
           axios.post(get_mongo_api(`students/add`), { _id: formValues._id }).then(res => {
             if (!res.data.success) {
               alert(res.data.message)
             }
           })
-        } else {
+        } else if (was_student && formValues.isStudent) {
+          let name = `${formValues.first_name} ${formValues.last_name}`
+          axios.post(get_mongo_api(`students/update/name`), {_id : formValues._id , name } ).then(res => {
+            if (!res.data.success) {
+              alert(res.data.message)
+            }
+          })
+        } else if (was_student && !formValues.isStudent){
           axios.delete(get_mongo_api(`students/${formValues._id}`)).then(res => {
             if (!res.data.success) {
               alert(res.data.message)
@@ -186,7 +199,7 @@ function AccountMenu({ handleSubmit, formValues, next_role, userDetails, navbar_
       if (!validForm) {
         alert(local_errors);
       } else {
-        httpPostRequestToUpdateUser(formValues)
+        httpPostRequestToUpdateUser(formValues, userDetails.isTeacher, userDetails.isStudent)
       }
     }
   }
